Extract skip_preload implementation to CommonOpts module
