+++ lib/permit_phoenix/live_view.ex
@@ -1389,28 +1389,7 @@ defmodule Permit.Phoenix.LiveView do
   end
 
   @doc false
-  def skip_preload(opts) do
-    cond do
-      # skip_preload option takes precedence
-      is_list(opts[:skip_preload]) ->
-        opts[:skip_preload]
-
-      # deprecated: if preload_actions is set, emit warning and convert to skip_preload
-      is_list(opts[:preload_actions]) ->
-        IO.warn(
-          "The :preload_actions option is deprecated. Use :skip_preload instead. " <>
-            "Actions not in skip_preload will automatically preload records.",
-          Macro.Env.stacktrace(__ENV__)
-        )
-
-        # Can't reliably convert preload_actions to skip_preload, so we return default
-        [:create, :new]
-
-      # default
-      true ->
-        [:create, :new]
-    end
-  end
+  defdelegate skip_preload(opts), to: Permit.Phoenix.CommonOpts
 
   @doc false
   def loader(resolution_context, opts) do
